Adjust size_t and uint64_t types
ngtcp2_vec_len now returns uint64_t and changes source code
accordingly.  Make sure that program compiles with 32bit compiler.

+++ lib/ngtcp2_conn.h
@@ -269,9 +269,9 @@ typedef struct ngtcp2_pktns {
       struct {
         /* ect0, ect1, ce are the ECN counts received in the latest
            ACK frame. */
-        size_t ect0;
-        size_t ect1;
-        size_t ce;
+        uint64_t ect0;
+        uint64_t ect1;
+        uint64_t ce;
       } ack;
     } ecn;
   } rx;
