+++ lib/ngtcp2_pkt.h
@@ -1125,7 +1125,7 @@ int ngtcp2_pkt_validate_ack(ngtcp2_ack *fr);
  * small to write STREAM frame, this function returns (size_t)-1.
  */
 size_t ngtcp2_pkt_stream_max_datalen(int64_t stream_id, uint64_t offset,
-                                     size_t len, size_t left);
+                                     uint64_t len, size_t left);
 
 /*
  * ngtcp2_pkt_crypto_max_datalen returns the maximum number of bytes
